fix: finish runtime fact consolidation
